+++ src/common/buffer/Types.ts
@@ -4,7 +4,8 @@
  */
 
 import type { IEvent } from '../Event';
-import type { ICircularList, ICharset } from '../Types';
+import type { ICircularList } from '../CircularList';
+import type { ICharset } from '../Types';
 import type { IDisposable } from '../Lifecycle';
 import type { UnderlineStyle } from './Constants';
